Split data loading unit tests to different files
